Move duplicated getPropertyName to util function
