+++ lib/rules/default-props-match-prop-types.js
@@ -9,6 +9,7 @@ const has = require('has');
 const Components = require('../util/Components');
 const variableUtil = require('../util/variable');
 const annotations = require('../util/annotations');
+const astUtil = require('../util/ast');
 
 // ------------------------------------------------------------------------------
 // Rule Definition
@@ -39,26 +40,12 @@ module.exports = {
     const propWrapperFunctions = new Set(context.settings.propWrapperFunctions || []);
 
     /**
-     * Get properties name
-     * @param {Object} node - Property.
-     * @returns {String} Property name.
-     */
-    function getPropertyName(node) {
-      if (node.key || ['MethodDefinition', 'Property'].indexOf(node.type) !== -1) {
-        return node.key.name;
-      } else if (node.type === 'MemberExpression') {
-        return node.property.name;
-      }
-      return '';
-    }
-
-    /**
      * Checks if the Identifier node passed in looks like a propTypes declaration.
      * @param   {ASTNode}  node The node to check. Must be an Identifier node.
      * @returns {Boolean}       `true` if the node is a propTypes declaration, `false` if not
      */
     function isPropTypesDeclaration(node) {
-      return getPropertyName(node) === 'propTypes';
+      return astUtil.getPropertyName(node) === 'propTypes';
     }
 
     /**
@@ -67,7 +54,7 @@ module.exports = {
      * @returns {Boolean}       `true` if the node is a defaultProps declaration, `false` if not
      */
     function isDefaultPropsDeclaration(node) {
-      const propName = getPropertyName(node);
+      const propName = astUtil.getPropertyName(node);
       return (propName === 'defaultProps' || propName === 'getDefaultProps');
     }
 
@@ -336,7 +323,7 @@ module.exports = {
     }
 
     function isPropTypeAnnotation(node) {
-      return (getPropertyName(node) === 'props' && !!node.typeAnnotation);
+      return (astUtil.getPropertyName(node) === 'props' && !!node.typeAnnotation);
     }
 
     function propFromName(propTypes, name) {
@@ -528,7 +515,7 @@ module.exports = {
           return;
         }
 
-        const propName = getPropertyName(node);
+        const propName = astUtil.getPropertyName(node);
         const isPropType = propName === 'propTypes';
         const isDefaultProp = propName === 'defaultProps' || propName === 'getDefaultProps';
