+++ lib/rules/no-redundant-should-component-update.js
@@ -4,6 +4,7 @@
 'use strict';
 
 const Components = require('../util/Components');
+const astUtil = require('../util/ast');
 
 function errorMessage(node) {
   return `${node} does not need shouldComponentUpdate when extending React.PureComponent.`;
@@ -24,18 +25,6 @@ module.exports = {
   },
 
   create: Components.detect((context, components, utils) => {
-    /**
-     * Get properties name
-     * @param {Object} node - Property.
-     * @returns {String} Property name.
-     */
-    function getPropertyName(node) {
-      if (node.key) {
-        return node.key.name;
-      }
-      return '';
-    }
-
     /**
      * Get properties for a given AST node
      * @param {ASTNode} node The AST node being checked.
@@ -59,7 +48,7 @@ module.exports = {
     function hasShouldComponentUpdate(node) {
       const properties = getComponentProperties(node);
       return properties.some(property => {
-        const name = getPropertyName(property);
+        const name = astUtil.getPropertyName(property);
         return name === 'shouldComponentUpdate';
       });
     }
