+++ lib/rules/prefer-stateless-function.js
@@ -9,6 +9,7 @@
 const has = require('has');
 const Components = require('../util/Components');
 const versionUtil = require('../util/version');
+const astUtil = require('../util/ast');
 
 // ------------------------------------------------------------------------------
 // Rule Definition
@@ -42,15 +43,6 @@ module.exports = {
     // --------------------------------------------------------------------------
     // Public
     // --------------------------------------------------------------------------
-
-    /**
-     * Get properties name
-     * @param {Object} node - Property.
-     * @returns {String} Property name.
-     */
-    function getPropertyName(node) {
-      return node.key.name;
-    }
 
     /**
      * Get properties for a given AST node
@@ -205,7 +197,7 @@ module.exports = {
     function hasOtherProperties(node) {
       const properties = getComponentProperties(node);
       return properties.some(property => {
-        const name = getPropertyName(property);
+        const name = astUtil.getPropertyName(property);
         const isDisplayName = name === 'displayName';
         const isPropTypes = name === 'propTypes' || name === 'props' && property.typeAnnotation;
         const contextTypes = name === 'contextTypes';
@@ -311,7 +303,7 @@ module.exports = {
         }
         // Ignore `props` and `context`
         const useThis = node.id.properties.some(property => {
-          const name = getPropertyName(property);
+          const name = astUtil.getPropertyName(property);
           return name !== 'props' && name !== 'context';
         });
         if (!useThis) {
