+++ lib/rules/require-default-props.js
@@ -8,6 +8,7 @@ const has = require('has');
 const Components = require('../util/Components');
 const variableUtil = require('../util/variable');
 const annotations = require('../util/annotations');
+const astUtil = require('../util/ast');
 
 const QUOTES_REGEX = /^["']|["']$/g;
 
@@ -30,26 +31,12 @@ module.exports = {
     const propWrapperFunctions = new Set(context.settings.propWrapperFunctions || []);
 
     /**
-     * Get properties name
-     * @param {Object} node - Property.
-     * @returns {String} Property name.
-     */
-    function getPropertyName(node) {
-      if (node.key || ['MethodDefinition', 'Property'].indexOf(node.type) !== -1) {
-        return node.key.name;
-      } else if (node.type === 'MemberExpression') {
-        return node.property.name;
-      }
-      return '';
-    }
-
-    /**
      * Checks if the Identifier node passed in looks like a propTypes declaration.
      * @param   {ASTNode}  node The node to check. Must be an Identifier node.
      * @returns {Boolean}       `true` if the node is a propTypes declaration, `false` if not
      */
     function isPropTypesDeclaration(node) {
-      return getPropertyName(node) === 'propTypes';
+      return astUtil.getPropertyName(node) === 'propTypes';
     }
 
     /**
@@ -58,7 +45,7 @@ module.exports = {
      * @returns {Boolean}       `true` if the node is a defaultProps declaration, `false` if not
      */
     function isDefaultPropsDeclaration(node) {
-      return (getPropertyName(node) === 'defaultProps' || getPropertyName(node) === 'getDefaultProps');
+      return (astUtil.getPropertyName(node) === 'defaultProps' || astUtil.getPropertyName(node) === 'getDefaultProps');
     }
 
     /**
@@ -306,7 +293,7 @@ module.exports = {
     }
 
     function isPropTypeAnnotation(node) {
-      return (getPropertyName(node) === 'props' && !!node.typeAnnotation);
+      return (astUtil.getPropertyName(node) === 'props' && !!node.typeAnnotation);
     }
 
     /**
@@ -483,8 +470,8 @@ module.exports = {
           return;
         }
 
-        const isPropType = getPropertyName(node) === 'propTypes';
-        const isDefaultProp = getPropertyName(node) === 'defaultProps' || getPropertyName(node) === 'getDefaultProps';
+        const isPropType = astUtil.getPropertyName(node) === 'propTypes';
+        const isDefaultProp = astUtil.getPropertyName(node) === 'defaultProps' || astUtil.getPropertyName(node) === 'getDefaultProps';
 
         if (!isPropType && !isDefaultProp) {
           return;
