+++ lib/rules/require-render-return.js
@@ -6,6 +6,7 @@
 
 const has = require('has');
 const Components = require('../util/Components');
+const astUtil = require('../util/ast');
 
 // ------------------------------------------------------------------------------
 // Rule Definition
@@ -49,23 +50,6 @@ module.exports = {
     }
 
     /**
-     * Get properties name
-     * @param {Object} node - Property.
-     * @returns {String} Property name.
-     */
-    function getPropertyName(node) {
-      // Special case for class properties
-      // (babel-eslint does not expose property name so we have to rely on tokens)
-      if (node.type === 'ClassProperty') {
-        const tokens = context.getFirstTokens(node, 2);
-        return tokens[1] && tokens[1].type === 'Identifier' ? tokens[1].value : tokens[0].value;
-      } else if (['MethodDefinition', 'Property'].indexOf(node.type) !== -1) {
-        return node.key.name;
-      }
-      return '';
-    }
-
-    /**
      * Check if a given AST node has a render method
      * @param {ASTNode} node The AST node being checked.
      * @returns {Boolean} True if there is a render method, false if not
@@ -73,7 +57,7 @@ module.exports = {
     function hasRenderMethod(node) {
       const properties = getComponentProperties(node);
       for (let i = 0, j = properties.length; i < j; i++) {
-        if (getPropertyName(properties[i]) !== 'render' || !properties[i].value) {
+        if (astUtil.getPropertyName(properties[i]) !== 'render' || !properties[i].value) {
           continue;
         }
         return /FunctionExpression$/.test(properties[i].value.type);
@@ -91,7 +75,7 @@ module.exports = {
           }
           if (
             !/(MethodDefinition|(Class)?Property)$/.test(ancestors[i].type) ||
-            getPropertyName(ancestors[i]) !== 'render' ||
+            astUtil.getPropertyName(ancestors[i]) !== 'render' ||
             depth > 1
           ) {
             continue;
@@ -101,7 +85,7 @@ module.exports = {
       },
 
       ArrowFunctionExpression: function(node) {
-        if (node.expression === false || getPropertyName(node.parent) !== 'render') {
+        if (node.expression === false || astUtil.getPropertyName(node.parent) !== 'render') {
           return;
         }
         markReturnStatementPresent(node);
