+++ lib/rules/sort-comp.js
@@ -8,6 +8,7 @@ const has = require('has');
 const util = require('util');
 
 const Components = require('../util/Components');
+const astUtil = require('../util/ast');
 
 /**
  * Get the methods order from the default config and the user config
@@ -203,7 +204,7 @@ module.exports = {
         return 'setter functions';
       }
 
-      return node.key.name;
+      return astUtil.getPropertyName(node);
     }
 
     /**
